rename google stackdriver to google cloud logging
