+++ homeassistant/components/google_cloud_logging/manifest.json
@@ -1,0 +1,13 @@
+{
+  "domain": "google_cloud_logging",
+  "name": "Google Cloud Logging",
+  "config_flow": false,
+  "documentation": "https://www.home-assistant.io/integrations/google_cloud_logging",
+  "requirements": [
+    "google-cloud-logging==1.14.0"
+  ],
+  "dependencies": [],
+  "codeowners": [
+    "@frog32"
+  ]
+}
