--- homeassistant/components/google_stackdriver/manifest.json
@@ -1,13 +1,0 @@
-{
-  "domain": "google_stackdriver",
-  "name": "Google Stackdriver",
-  "config_flow": false,
-  "documentation": "https://www.home-assistant.io/integrations/google_stackdriver",
-  "requirements": [
-    "google-cloud-logging==1.14.0"
-  ],
-  "dependencies": [],
-  "codeowners": [
-    "@frog32"
-  ]
-}
