+++ requirements_all.txt
@@ -617,7 +617,7 @@ gntp==1.0.3
 # homeassistant.components.google
 google-api-python-client==1.6.4
 
-# homeassistant.components.google_stackdriver
+# homeassistant.components.google_cloud_logging
 google-cloud-logging==1.14.0
 
 # homeassistant.components.google_pubsub
