+++ requirements_test_all.txt
@@ -238,7 +238,7 @@ glances_api==0.2.0
 # homeassistant.components.google
 google-api-python-client==1.6.4
 
-# homeassistant.components.google_stackdriver
+# homeassistant.components.google_cloud_logging
 google-cloud-logging==1.14.0
 
 # homeassistant.components.google_pubsub
