+++ tests/components/google_cloud_logging/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the Google Cloud Logging integration."""
