--- tests/components/google_stackdriver/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the Google Stackdriver integration."""
